requirements.txt
pandas~=2.2.3
rapidfuzz~=3.12.2
geopandas~=1.0.1
geopy~=2.4.1
tqdm~=4.67.1
numpy~=2.0.2
shapely~=2.0.7
PyYAML~=6.0.2
recordlinkage~=0.16